+++ src/main/java/com/dace/dmgr/combat/character/vellion/Vellion.java
@@ -170,7 +170,7 @@ public final class Vellion extends Controller {
         super.onGiveHeal(provider, target, amount);
 
         if (provider != target && target instanceof CombatUser)
-            provider.addScore("치유", (double) (HEAL_SCORE * amount) / target.getDamageModule().getMaxHealth());
+            provider.addScore("치유", HEAL_SCORE * amount / target.getDamageModule().getMaxHealth());
 
         return true;
     }
